==> 54.parse.AFandAFD.rotations.Rmd <==
---
title: "54.parse.AFandAFD.rotations"
author: "Daniele Filiault"
date: "7/21/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(glue)
library(ggplot2)
library(viridis)
library(RColorBrewer)
library(grid)
library(ggpubr)
```

## Introduction

Parsing rotations done in script(s) 53 to do genome rotation to get empirical distributions of Kologorov-Smirnov test values obtained in script 52

```{r load data}
scans <- c("n.scan.rot", "s.scan.rot", "n.gea.rot", "s.gea.rot","n.fit.rot", "s.fit.rot", "n.fit2.rot", "s.fit2.rot")
vars <- c("n.sweep.rot", "s.sweep.rot", "n.gea.rot", "s.gea.rot","n.fit.rot", "s.fit.rot", "n.fit.rot", "s.fit.rot")
all.dat <- as.list(1:length(scans))
for(up in 1:length(scans)){
  print(up)
  up.scan <- scans[up]
  print(up.scan)
  up.var <- vars[up]
  #print(up.var)
  up.files <- glue('./data/53.data/{up.scan}.{1:25}.Rdat')
  for(up.f in up.files){
    load(up.f)
    up.dat <- do.call(rbind, get(up.var))
    if(exists("out.dat")){out.dat <- rbind(out.dat, up.dat)
    }else{out.dat <- up.dat}
  }
  out.dat <- unique(out.dat)
  out.dat$scan.type <- up.scan
  all.dat[[up]] <- out.dat
  rm(out.dat)
}
all.dat <- do.call(rbind, all.dat)

all.dat$scan.type <-  substring(all.dat$scan.type,3)
all.dat$exp.site <- do.call(rbind,strsplit(all.dat$exp, "_"))[,1]
all.dat$year <- do.call(rbind,strsplit(all.dat$exp, "_"))[,2]
all.dat$scan.type <- gsub("fit2.rot", "fit.rot", all.dat$scan.type) ### combine 2 rounds of fitness scans

```

## 01. plot distribution of KS values for AF/AFD and betas pairwise


```{r AF beta KS plot}
scannames <- unique(all.dat$scan)

for(up.scan in scannames){
  
  up.af.plotfile <- glue('./figures/54.figures/{up.scan}.AF.beta.jpg')
  up.afd.plotfile <- glue('./figures/54.figures/{up.scan}.AFD.beta.jpg')

  scan.af.plot <- ggplot(dat=all.dat[all.dat$rotation!="observed" & all.dat$scan==up.scan, ], aes(x=beta.statistic, y=af.statistic)) + 
    geom_point(alpha=0.09) + 
    facet_grid(year~exp.site) + 
    geom_point(data=all.dat[all.dat$rotation=="observed" & all.dat$scan==up.scan,], color="red", size=3) +
    ggtitle(up.scan) +
    theme_bw() +
    xlab("betas KS statistic") +
    ylab("allele frequencey KS statistic")
  ggsave(scan.af.plot, file=up.af.plotfile, width=8, height=6)

  scan.afd.plot <- ggplot(dat=all.dat[all.dat$rotation!="observed" & all.dat$scan==up.scan, ], aes(x=beta.statistic, y=afd.statistic)) + 
    geom_point(alpha=0.09) + 
    facet_grid(year~exp.site) + 
    geom_point(data=all.dat[all.dat$rotation=="observed" & all.dat$scan==up.scan,], color="red", size=3) +
    ggtitle(up.scan) +
    xlab("beta KS statistic") +
    ylab("allele frequencey difference KS statistic") +
    theme_bw()
  ggsave(scan.afd.plot, file=up.afd.plotfile, width=8, height=6)

}


```

## 02. get number of rotations with smaller KS than observed for all exp/year by scan combinations

```{r get nrotations smaller than observed}
expnames <- unique(all.dat$exp)
scannames <- unique(all.dat$scan)

scan.pvals <- as.list(1:length(expnames))
for(up in 1:length(expnames)){
  up.exp <- expnames[up]
  oe.dat <- sapply(1:length(scannames), function(up.s){
    up.scan <- scannames[up.s]
    up.dat <- all.dat[all.dat$scan==up.scan & all.dat$exp==up.exp,]
    n.af <- sum(up.dat$af.statistic < up.dat[up.dat$rotation=="observed","af.statistic"])
    n.afd <- sum(up.dat$afd.statistic < up.dat[up.dat$rotation=="observed","afd.statistic"])
    n.beta <- sum(up.dat$beta.statistic < up.dat[up.dat$rotation=="observed","beta.statistic"])
    out.dat <- unique(up.dat[,colnames(up.dat)%in%c("exp","scan","scan.type","exp.site", "year")])
    out.dat$n.af <- n.af
    out.dat$n.afd <- n.afd
    out.dat$n.beta <- n.beta
    return(out.dat)
      })
  oe.dat <- t(oe.dat)
  scan.pvals[[up]] <- oe.dat
}

scan.pvals <- do.call(rbind, scan.pvals)
scan.pvals <- apply(scan.pvals, 2, as.character)
scan.pvals <- as.data.frame(scan.pvals, stringsAsFactors=FALSE)
scan.pvals$scan.type <- gsub(".rot", "", scan.pvals$scan.type)
scan.pvals$n.af <- as.numeric(scan.pvals$n.af)
scan.pvals$n.afd <- as.numeric(scan.pvals$n.afd)
scan.pvals$n.beta <- as.numeric(scan.pvals$n.beta)

```

## 03. plot these "empirical pvalues" by experiment, one plot per scan type for clarity

```{r generate AF/AFD and beta pval plots}
types <- unique(scan.pvals$scan.type)
scan.pvals$exp.site <- gsub("ADA", "NA", scan.pvals$exp.site)
scan.pvals$exp.site <- gsub("RAM", "NM", scan.pvals$exp.site)
scan.pvals$exp.site <- gsub("RAT", "SR", scan.pvals$exp.site)
scan.pvals$exp.site <- gsub("ULL", "SU", scan.pvals$exp.site)

type.af.plots <- as.list(1:3)
type.afd.plots <- as.list(1:3)

types.long <- c("A  selection scan", "B  genotype environment association", "C  field fitness")
pals <- c("Set2","Accent","Set1")

for(up in 1:length(types)){
  up.t <- types[up]
  up.title <- types.long[up]
  up.pal <- pals[up] 
  
  af.pplot <- ggplot(data=scan.pvals[scan.pvals$scan.type==up.t,], aes(x=n.beta, y=n.af, color=scan)) +
    geom_point() +
    facet_grid(year~exp.site) +
    scale_color_brewer(palette=up.pal, name="scan name") +
    #scale_colour_viridis_d(name="scan name") +
    xlab("number of rotations with KS less than observed (beta)") +
    ylab("number of rotations with \nKS less than observed \n(allele frequency)") +
    theme_bw() +
    ggtitle(up.title) +
    theme(axis.text.x = element_text(angle = 45, hjust=1)) +
    xlim(0,1000) +
    ylim(0,1000)
  type.af.plots[[up]] <- af.pplot
  
  afd.pplot <- ggplot(data=scan.pvals[scan.pvals$scan.type==up.t,], aes(x=n.beta, y=n.afd, color=scan)) +
    geom_point() +
    facet_grid(year~exp.site) +
    scale_color_brewer(palette=up.pal, name="scan name") +
    xlab("number of rotations with KS less than observed (beta)") +
    ylab("number of rotations with \nKS less than observed \n(allele frequency difference)") +
    theme_bw() +
    ggtitle(up.title)+
    theme(axis.text.x = element_text(angle = 45, hjust=1)) +
    xlim(0,1000) +
    ylim(0,1000)
  type.afd.plots[[up]] <- afd.pplot
  }

```

## 04. generate single figures from figures in 03
```{r consolidate AF figures}

af.figure <- ggarrange(type.af.plots[[1]] + rremove("ylab") + rremove("xlab"), type.af.plots[[2]] + rremove("ylab") + rremove("xlab"), type.af.plots[[3]] + rremove("ylab") + rremove("xlab"), # remove axis labels from plots
                    labels = NULL,
                    ncol = 1, nrow = 3,
                    common.legend = FALSE,
                    align = "hv", 
                    font.label = list(size = 10, color = "black", face = "bold", family = NULL, position = "top"))

af.figure <- annotate_figure(af.figure, left = textGrob("number of rotations with KS less than observed (allele frequency)", rot = 90, vjust = 1, gp = gpar(cex = 1.1)), bottom = textGrob("number of rotations with KS less than observed (beta)                     ", gp = gpar(cex = 1.1)))

ggsave(af.figure, file="./figures/54.figures/af.beta.all.jpg", width=7, height=9)

```

```{r consolidate AFD figures}

afd.figure <- ggarrange(type.afd.plots[[1]] + rremove("ylab") + rremove("xlab"), type.afd.plots[[2]] + rremove("ylab") + rremove("xlab"), type.afd.plots[[3]] + rremove("ylab") + rremove("xlab"), # remove axis labels from plots
                    labels = NULL,
                    ncol = 1, nrow = 3,
                    common.legend = FALSE,
                    align = "hv", 
                    font.label = list(size = 10, color = "black", face = "bold", family = NULL, position = "top"))

afd.figure <- annotate_figure(afd.figure, left = textGrob("number of rotations with KS less than observed (allele frequency difference)", rot = 90, vjust = 1, gp = gpar(cex = 1.1)), bottom = textGrob("number of rotations with KS less than observed (beta)                     ", gp = gpar(cex = 1.1)))

ggsave(afd.figure, file="./figures/54.figures/afd.beta.all.jpg", width=7, height=9)

```
